add:jsファイルの読み込み成功→親カテゴリー選択で子カテゴリーのデータ取得
